Moved to Uint16, changed NOT_ALLOWED to RESOURCE_ERROR and redesigned test case

+++ src/lavinmq/config.cr
@@ -66,7 +66,7 @@ module LavinMQ
     property max_deleted_definitions = 8192         # number of deleted queues, unbinds etc that compacts the definitions file
     property consumer_timeout : UInt64? = nil
     property consumer_timeout_loop_interval = 60 # seconds
-    property consumer_max_per_channel = UInt64::MAX
+    property consumer_max_per_channel = UInt16::MAX
     property default_consumer_prefetch = UInt16::MAX
     property yield_each_received_bytes = 131_072    # max number of bytes to read from a client connection without letting other tasks in the server do any work
     property yield_each_delivered_bytes = 1_048_576 # max number of bytes sent to a client without tending to other tasks in the server
